+++ packages/gatsby/src/utils/webpack-utils.ts
@@ -276,7 +276,6 @@ export const createWebpackUtils = (
         options: {
           stage,
           reactRuntime: jsxRuntimeExists ? `automatic` : `classic`,
-          // TODO add proper cache keys
           cacheDirectory: path.join(
             program.directory,
             `.cache`,
@@ -284,6 +283,7 @@ export const createWebpackUtils = (
             `babel`
           ),
           ...options,
+          rootDir: program.directory,
         },
         loader: require.resolve(`./babel-loader`),
       }
@@ -292,7 +292,6 @@ export const createWebpackUtils = (
     dependencies: options => {
       return {
         options: {
-          // TODO add proper cache keys
           cacheDirectory: path.join(
             program.directory,
             `.cache`,
@@ -400,9 +399,11 @@ export const createWebpackUtils = (
         // debugger to show the original code. Instead, the code
         // being evaluated would be much more helpful.
         sourceMaps: false,
-        cacheIdentifier: `${stage}---gatsby-dependencies@${
-          require(`babel-preset-gatsby/package.json`).version
-        }`,
+
+        cacheIdentifier: JSON.stringify({
+          browerslist: supportedBrowsers,
+          gatsbyPreset: require(`babel-preset-gatsby/package.json`).version,
+        }),
       }
 
       // TODO REMOVE IN V3