+++ packages/gatsby-plugin-mdx/CHANGELOG.md
@@ -3,6 +3,12 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.2.41](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.2.40...gatsby-plugin-mdx@1.2.41) (2020-09-24)
+
+### Bug Fixes
+
+- **gatsby-plugin-mdx:** Fix not passing file path to remark plugins ([#26914](https://github.com/gatsbyjs/gatsby/issues/26914)) ([5d39594](https://github.com/gatsbyjs/gatsby/commit/5d39594c1d9b3562b318b5d299905c0078d74557))
+
 ## [1.2.40](https://github.com/gatsbyjs/gatsby/compare/gatsby-plugin-mdx@1.2.39...gatsby-plugin-mdx@1.2.40) (2020-09-15)
 
 **Note:** Version bump only for package gatsby-plugin-mdx